+++ packages/gatsby/src/redux/types.ts
@@ -4,6 +4,7 @@ import { DocumentNode, GraphQLSchema } from "graphql"
 import { SchemaComposer } from "graphql-compose"
 import { IGatsbyCLIState } from "gatsby-cli/src/reporter/redux/types"
 import { InternalJobInterface, JobResultInterface } from "../utils/jobs-manager"
+import { ITypeMetadata } from "../schema/infer/inference-metadata"
 
 type SystemPath = string
 type Identifier = string
@@ -75,6 +76,7 @@ export interface IGatsbyNode {
   }
   __gatsby_resolved: any // TODO
   [key: string]: unknown
+  fields: string[]
 }
 
 export interface IGatsbyPlugin {
@@ -215,24 +217,24 @@ export interface IGatsbyState {
     }
   }
   schemaCustomization: {
-    composer: SchemaComposer<any>
-    context: {} // TODO
-    fieldExtensions: {} // TODO
-    printConfig: any // TODO
-    thridPartySchemas: any[] // TODO
-    types: any[] // TODO
+    composer: null | SchemaComposer<any>
+    context: Record<string, any>
+    fieldExtensions: GraphQLFieldExtensionDefinition
+    printConfig: {
+      path?: string
+      include?: { types?: Array<string>; plugins?: Array<string> }
+      exclude?: { types?: Array<string>; plugins?: Array<string> }
+      withFieldTypes?: boolean
+    } | null
+    thirdPartySchemas: GraphQLSchema[]
+    types: (string | { typeOrTypeDef: DocumentNode; plugin: IGatsbyPlugin })[]
   }
   themes: any // TODO
   logs: IGatsbyCLIState
   inferenceMetadata: {
     step: string // TODO make enum or union
     typeMap: {
-      [key: string]: {
-        ignoredFields: Set<string>
-        total: number
-        dirty: boolean
-        fieldMap: any // TODO
-      }
+      [key: string]: ITypeMetadata
     }
   }
   pageDataStats: Map<SystemPath, number>
@@ -301,6 +303,12 @@ export type ActionsUnion =
   | IAddPendingPageDataWriteAction
   | IAddPendingTemplateDataWriteAction
   | IClearPendingPageDataWritesAction
+  | ICreateResolverContext
+  | IClearSchemaCustomizationAction
+  | ISetSchemaComposerAction
+  | IStartIncrementalInferenceAction
+  | IBuildTypeMetadataAction
+  | IDisableTypeInferenceAction
 
 interface ISetBabelPluginAction {
   type: `SET_BABEL_PLUGIN`
@@ -513,6 +521,15 @@ export interface ICreateResolverContext {
     | { [camelCasedPluginNameWithoutPrefix: string]: IGatsbyPluginContext }
 }
 
+interface IClearSchemaCustomizationAction {
+  type: `CLEAR_SCHEMA_CUSTOMIZATION`
+}
+
+interface ISetSchemaComposerAction {
+  type: `SET_SCHEMA_COMPOSER`
+  payload: SchemaComposer<any>
+}
+
 export interface ICreatePageAction {
   type: `CREATE_PAGE`
   payload: IGatsbyPage
@@ -631,11 +648,13 @@ export interface ISetSiteConfig {
 export interface ICreateNodeAction {
   type: `CREATE_NODE`
   payload: IGatsbyNode
+  oldNode?: IGatsbyNode
 }
 
 export interface IAddFieldToNodeAction {
   type: `ADD_FIELD_TO_NODE`
   payload: IGatsbyNode
+  addedField: string
 }
 
 export interface IAddChildNodeToParentNodeAction {
@@ -651,6 +670,7 @@ export interface IDeleteNodeAction {
 export interface IDeleteNodesAction {
   type: `DELETE_NODES`
   payload: Identifier[]
+  fullNodes: IGatsbyNode[]
 }
 
 export interface ISetSiteFlattenedPluginsAction {
@@ -677,4 +697,21 @@ export interface IAddPageDataStatsAction {
 export interface ITouchNodeAction {
   type: `TOUCH_NODE`
   payload: Identifier
+}
+
+interface IStartIncrementalInferenceAction {
+  type: `START_INCREMENTAL_INFERENCE`
+}
+
+interface IBuildTypeMetadataAction {
+  type: `BUILD_TYPE_METADATA`
+  payload: {
+    nodes: IGatsbyNode[]
+    typeName: string
+  }
+}
+
+interface IDisableTypeInferenceAction {
+  type: `DISABLE_TYPE_INFERENCE`
+  payload: string[]
 }