+++ packages/gatsby-plugin-google-gtag/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-gtag@2.1.3...gatsby-plugin-google-gtag@2.1.4) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
 ## [2.1.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-gtag@2.1.2...gatsby-plugin-google-gtag@2.1.3) (2020-06-02)
 
 **Note:** Version bump only for package gatsby-plugin-google-gtag